refactor(chain): better test case
